% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_request.R
\name{mr_build_request}
\alias{mr_build_request}
\title{Build a request for SIPRI data}
\usage{
mr_build_request()
}
\value{
An \code{httr2} request object configured with the specified parameters.
}
\description{
Constructs an HTTP request for the SIPRI data export.
}
\keyword{internal}
