% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipri_request.R
\name{sipri_build_request}
\alias{sipri_build_request}
\title{Build a request for SIPRI data}
\usage{
sipri_build_request(verbose = FALSE)
}
\value{
An \code{httr2} request object configured with the specified parameters.
}
\description{
Constructs an HTTP request for the SIPRI data export.
}
\keyword{internal}
