% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipri_get_data.R
\name{sipri_get_data}
\alias{sipri_get_data}
\title{Get data from SIPRI Military Expenditure (Milex) Database}
\usage{
sipri_get_data(
  indicator = "constantUSD",
  verbose = FALSE,
  footnotes = FALSE,
  cache = TRUE
)
}
\arguments{
\item{indicator}{Character vector. The type of data to download. Must be one of:
"constantUSD", "currentUSD", "shareOfGDP", "shareGovt", "regionalTotals",
"currencyFY", "currencyCY", "perCapita", or "all".
Default is "constantUSD".}

\item{verbose}{Logical. If TRUE, prints additional details about the query
process. Default is FALSE.}

\item{footnotes}{Logical. If TRUE, joins footnotes to the result table.
Default is FALSE.}

\item{cache}{Logical. If TRUE, caches the response and processed
data to avoid re-querying. Default is TRUE.}
}
\value{
A tibble containing the requested data.
}
\description{
The \code{sipri_get_data} function retrieves military expenditure data from the
SIPRI Milex database. It provides multiple indicators including constant USD,
current USD, share of GDP, and more. You can also join footnotes to the
results and enable caching for faster queries.
}
\examples{
\dontrun{
# Retrieve military expenditure data in constant 2022 USD
milex_data <- sipri_get_data(indicator = "constantUSD",
    verbose = TRUE,
    footnotes = TRUE)

# Retrieve regional totals without caching
regional_totals <- sipri_get_data(indicator = "regionalTotals", cache = FALSE)
}

}
