% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sipri_get_footnotes}
\alias{sipri_get_footnotes}
\title{Extract and format footnotes from an Excel file}
\usage{
sipri_get_footnotes(file_path, notes)
}
\arguments{
\item{file_path}{A character string specifying the path to the Excel file containing footnotes.}

\item{notes}{A character vector of notes to extract and format.}
}
\value{
A data frame with two columns: \code{notes}, containing the original notes, and \code{footnote}, containing the formatted footnotes.
}
\description{
Reads footnotes from a specified Excel file and formats them according to the provided notes.
}
\keyword{internal}
