% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sipri_regions}
\alias{sipri_regions}
\title{Regions data by SIPRI}
\format{
\subsection{\code{sipri_regions}}{

A data frame with 7,240 rows and 60 columns:
\describe{
\item{region}{Region name}
\item{iso3c}{3 letter ISO country codes}
\item{country}{Country name}
}
}
}
\source{
\url{https://milex.sipri.org/sipri}
}
\usage{
sipri_regions
}
\description{
ISO3 Country codes and regional codes used by SIPRI for matching to the data.
}
\keyword{datasets}
