% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sipri_get_data.R
\name{sipri_get_data}
\alias{sipri_get_data}
\title{Get data from SIPRI Milex database}
\usage{
sipri_get_data(
  indicator = "constantUSD",
  verbose = FALSE,
  footnotes = FALSE,
  cache = TRUE
)
}
\arguments{
\item{indicator}{The type of data to download.}

\item{verbose}{Whether to be verbose.}

\item{footnotes}{Whether to join footnotes to the results table.}

\item{cache}{Whether to cache response from SIPRI and processed data.}
}
\value{
A tibble.
}
\description{
Get data from SIPRI Milex database
}
